Feat: 프로필 이미지 input 컴포넌트화 #28
