+++ src/components/MyPage/ProfileMenuBox.module.css
@@ -22,39 +22,6 @@
   }
 }
 
-.profileContainer {
-  width: 100%;
-  height: 16rem;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.profileContainer > .imgInput {
-  display: none;
-}
-
-.profileImg {
-  border-radius: 16rem;
-  box-shadow: 0px 4px 16px 0px rgba(0, 0, 0, 0.08);
-}
-
-.editIcon {
-  position: absolute;
-  bottom: 0;
-  right: 8rem;
-  cursor: pointer;
-
-  @media tablet {
-    right: 2rem;
-  }
-
-  @media mobile {
-    right: 7rem;
-  }
-}
-
 .memuContainer {
   display: flex;
   flex-direction: column;
