+++ src/components/MyPage/ProfileMenuBox.tsx
@@ -12,6 +12,7 @@ import LogoImg from '#/images/img-kakao.png';
 import Image from 'next/image';
 import EditIcon from '#/icons/icon-edit.svg';
 import { useRouter } from 'next/router';
+import ProfileInput from '@/components/MyPage/ProfileInput';
 
 const MENU_LIST = [
   {
@@ -43,49 +44,15 @@ const MENU_LIST = [
 function ProfileMenuBox() {
   const router = useRouter();
   const [isSelected, setIsSelected] = useState('내 정보');
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const [imageSrc, setImageSrc] = useState<string | null>(null);
 
   const handleMenuItem = (item: { title: string; link: string }) => {
     setIsSelected(item.title);
     router.push(item.link);
   };
 
-  const onUpload = async (e: ChangeEvent<HTMLInputElement>) => {
-    if (e.target && e.target.files) {
-      const targetFiles = e.target.files[0];
-      const selectedFiles = URL.createObjectURL(targetFiles);
-      setImageSrc(selectedFiles);
-      handlePostProfile(targetFiles);
-    }
-  };
-
-  const handlePostProfile = (imgUrl: File) => {
-    const imgFormData = new FormData();
-    imgFormData.append('image', imgUrl);
-    //TODO 이미지 url 생성 api 연동, imgFormData 넘겨주기
-  };
-
-  const handleUploadImg = useCallback(() => {
-    if (!inputRef.current) {
-      return;
-    }
-    inputRef.current.click();
-  }, []);
-
   return (
     <div className={styles.profileBoxContainer}>
-      <div className={styles.profileContainer}>
-        <Image
-          src={imageSrc ? imageSrc : LogoImg}
-          className={styles.profileImg}
-          alt="profileImg"
-          width={160}
-          height={160}
-        />
-        {isSelected === '내 정보' && <EditIcon className={styles.editIcon} onClick={handleUploadImg} />}
-        <input type="file" accept="image/*" onChange={(e) => onUpload(e)} ref={inputRef} className={styles.imgInput} />
-      </div>
+      <ProfileInput isProfileBox={true} />
       <div className={styles.memuContainer}>
         {MENU_LIST.map((e, index) => {
           return (
