+++ src/components/MyPage/profileInput.module.css
@@ -1,0 +1,38 @@
+@values tablet, mobile from '@/styles/breakpoints.css';
+
+.profileContainer {
+  display: none;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 16rem;
+  position: relative;
+
+  @media mobile {
+    display: flex;
+  }
+}
+
+.display {
+  display: flex;
+}
+
+.profileContainer > .imgInput {
+  display: none;
+}
+
+.profileImg {
+  border-radius: 16rem;
+  box-shadow: 0px 4px 16px 0px rgba(0, 0, 0, 0.08);
+}
+
+.editIcon {
+  position: absolute;
+  bottom: 0;
+  right: 8rem;
+  cursor: pointer;
+
+  @media tablet {
+    right: 2rem;
+  }
+}
