+++ src/pages/mypage/index.tsx
@@ -1,13 +1,10 @@
 import MyPageLayout from '@/components/MyPage/MyPageLayout';
-import { ChangeEvent, ReactElement, useCallback, useRef, useState } from 'react';
+import { ReactElement } from 'react';
 import styles from './MyPage.module.css';
 import { FieldValues, useForm } from 'react-hook-form';
 import Input from '@/components/Input/Input';
 import { passwordCheck } from '@/utils/passwordCheck';
-import Image from 'next/image';
-import LogoImg from '#/images/img-kakao.png';
-import EditIcon from '#/icons/icon-edit.svg';
-import profileStyles from '@/components/MyPage/ProfileMenuBox.module.css';
+import ProfileInput from '@/components/MyPage/ProfileInput';
 function MyPage() {
   const methods = useForm<FieldValues>({
     mode: 'onTouched',
@@ -17,6 +14,7 @@ function MyPage() {
       email: 'test@test.com',
       password: '',
       passwordCheck: '',
+      profileImageUrl: '',
     },
   });
 
@@ -30,30 +28,6 @@ function MyPage() {
 
     console.log(data);
   };
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const [imageSrc, setImageSrc] = useState<string | null>(null);
-
-  const onUpload = async (e: ChangeEvent<HTMLInputElement>) => {
-    if (e.target && e.target.files) {
-      const targetFiles = e.target.files[0];
-      const selectedFiles = URL.createObjectURL(targetFiles);
-      setImageSrc(selectedFiles);
-      handlePostProfile(targetFiles);
-    }
-  };
-
-  const handlePostProfile = (imgUrl: File) => {
-    const imgFormData = new FormData();
-    imgFormData.append('image', imgUrl);
-    //TODO 이미지 url 생성 api 연동, imgFormData 넘겨주기
-  };
-
-  const handleUploadImg = useCallback(() => {
-    if (!inputRef.current) {
-      return;
-    }
-    inputRef.current.click();
-  }, []);
 
   return (
     <div className={styles.myPageContainer}>
@@ -65,23 +39,7 @@ function MyPage() {
           </button>
         </div>
 
-        <div className={profileStyles.profileContainer}>
-          <Image
-            src={imageSrc ? imageSrc : LogoImg}
-            className={profileStyles.profileImg}
-            alt="profileImg"
-            width={160}
-            height={160}
-          />
-          <EditIcon className={profileStyles.editIcon} onClick={handleUploadImg} />
-          <input
-            type="file"
-            accept="image/*"
-            onChange={(e) => onUpload(e)}
-            ref={inputRef}
-            className={profileStyles.imgInput}
-          />
-        </div>
+        <ProfileInput isProfileBox={false} />
       </div>
       <form onSubmit={handleSubmit(handleOnSubmit)} className={styles.formContainer} id="mypageForm">
         <Input
